* Clang compilation fix in AsmJit (didn't even have time to push my patches here before they were pushed upstream, only took 10 minutes from my bug report and patch for it to be pushed upstream! impressive turnaround). * Clang compilation workaround in type traits (some unsupported traits).

+++ src/asmjit/asmjit/src/asmjit/x86/x86compilerfunc.cpp
@@ -949,7 +949,7 @@ CompilerItem* X86CompilerFuncRet::translate(CompilerContext& cc) ASMJIT_NOTHROW
   uint32_t retValType = getFunc()->getDecl()->getReturnType();
   uint32_t i;
 
-  switch ((int)retValType)
+  switch (retValType)
   {
     case kX86VarTypeGpd:
     case kX86VarTypeGpq:
